Make LP Demo Results Arc10 readable
Randal downloaded 2.0.0, opened the file on his computer, and saved it to all verisons of Arc 10.x.  john was able to open it on his.  John did a quick fix and hard pasted this into the folder from the web and will zip that up and submitted to website as v2.0.1 unless he hears otherwise (2.0.01 maybe?)