Issue #89 Toolbox update
trying that last commit again:
Improved the parameter name for the new parameter, and the related parameters for the Y axis. Improved the help dialog for the new parameter. Improved the help dialog for the associated parameters.  This time the toolbox should load. I hope all the relative paths and the multitool still works.